backend/requirements.txt
# ==========================================
# AI Trader v2 – Backend Dependencies
# ==========================================
# Requires Python >= 3.11
# Install: uv pip install -r requirements.txt

# --- Framework ---
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.30.0,<1.0
python-multipart>=0.0.9,<1.0

# --- Database ---
sqlalchemy[asyncio]>=2.0.30,<3.0
asyncpg>=0.29.0,<1.0
alembic>=1.13.0,<2.0
psycopg2-binary>=2.9.9,<3.0

# --- Validation & Settings ---
pydantic[email]>=2.7.0,<3.0
pydantic-settings>=2.3.0,<3.0

# --- Auth ---
python-jose[cryptography]>=3.3.0,<4.0
passlib[bcrypt]>=1.7.4,<2.0

# --- Redis ---
redis>=5.0.0,<6.0

# --- Background Tasks ---
celery>=5.4.0,<6.0

# --- HTTP Client ---
httpx>=0.27.0,<1.0

# --- Logging ---
structlog>=24.1.0,<25.0
python-json-logger>=2.0.0,<3.0

# --- Encryption ---
cryptography>=42.0.0,<44.0

# --- Env ---
python-dotenv>=1.0.0,<2.0

# --- WebSocket ---
websockets>=12.0,<14.0
